Create seperate compontent for Hero
